Mandatory part is done

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: arabiai <arabiai@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/12 13:06:21 by arabiai           #+#    #+#             */
-/*   Updated: 2022/11/12 19:45:34 by arabiai          ###   ########.fr       */
+/*   Updated: 2022/11/14 13:07:32 by arabiai          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 int     ft_printf(const char *string_format, ...);
 void	ft_putchar_fd_count(char c, int fd, int *count);
 void	ft_putnbr_fd_count(int n, int fd, int *count);
-void	ft_putnbr_fd_hexa_count(unsigned int nb, int fd, int *count, char c);
+void    ft_putnbr_fd_hexa_count(size_t nb, int fd, int *count, char c);
 void	ft_putnbr_fd_uns_count(unsigned int n, int fd, int *count);
 void	ft_putstr_fd_count(char *s, int fd, int *count);
